Updated openstack/openstack Project: openstack/nova  0ca64a8690ac9de720131b522d29dcaed824eb8d
Handle version for PUT and POST in PlacementFixture

The SchedulerReportClient allows a "version" keyword argument to be
specified for its get/put/post methods and several functions pass a
version because they are sending data supported only in specific
microversions of the placement API.

The PlacementFixture has to mock the SchedulerReportClient
get/put/post/delete methods to swap in a Keystone session with
auth=None. The fake methods for PUT and POST are missing handling of
the "version" keyword argument, causing intermittent "400 Bad Request"
failures in the functional test jobs when unexpected parameters are
passed to placement microversion 1.0 instead of the intended
microversion.

Closes-Bug: #1701129

Change-Id: I681712ac37f732c7803c68f6c7d1eae9f2877d3d