Updated openstack/openstack Project: openstack/governance  adfd6f7e9bb3df269b2bcdeae6a5469bdc6b033f
Add logstash-filters as an Infra deliverable

The openstack-infra/logstash-filters repository was forked from
openstack-infra/system-config a couple years ago in
Ib6139fdf7d11f5bab9745590d2f93f57672b8e99 but not added as an
Infrastructure team deliverable at the time. Since it is relied on
by our logstash indexing processes and is mainly attended to by
Infrastructure contributors and reviewers, it makes sense to include
within the team's scope.

Change-Id: If05d0fa3eff59fd32d4fd6c7ab4e1f923de5a030