Updated openstack/openstack Project: openstack/tripleo-heat-templates  0de13ab13d327eebe307df36c9d02f9792521fdb
Add a docker-ha.yaml environment file for containerized HA deployments

This adds a docker-ha.yaml that can be passed to the deployment
environments in order to get a containerized HA deployment.
Until we make the containerized deplyment the default the operator
must first include docker.yaml and *then* docker-ha.yaml in order
to get a containerized overcloud with an HA control plane.

We also make sure that the ClusterCheck service is set to None
by default and is part of the Controller roles.

Change-Id: I13204d70aad8dfeaf2bcf2ae30a1bb4715167659