Updated openstack/openstack Project: openstack/nova  c140d7d708c22bd54e5020567ddd02ee392b00aa
Ensure the JSON-Schema covers the legacy v2 API

The legacy v2 API compatible mode support the protection of JSON-Schema.
The input body will be validated with JSON-Schema, and the extra invalid
parameters will be filtered out of the input body instead of return
HTTPBadRequest 400.

But some of API are missing that protection, the JSON-Schema validation
was limited to the v2.1 API. This patch ensures those schema covers the
legacy v2 API.

Change-Id: Ie165b2a79efd56a299d2d4ebe40a6904a340414f
Closes-Bug: #1701451