Updated openstack/openstack Project: openstack/cliff  cc02cc85f277c57d663edc7ebdb867a74907132e
add hook for manipulating the argument parser

Update Commands to load a separate set of extensions to be used as
"hooks," triggered at different points in the processing of the
command. Start with a hook that is given access to the argument parser
for the command so it can modify it.

Change-Id: I0785548fd36a61cda616921a4a21be3f67701300
Signed-off-by: Doug Hellmann <doug@doughellmann.com>