Updated openstack/openstack Project: openstack/horizon  ed941298e71c4c95435da4668803f0a1ff4e2a47
doc: Move settings information to configuration/ dir

The new standard document structure suggests to have a dedicated dir
named configuration/. It will be linked from docs.o.o directly,
so it would be better to follow the guideline.

This commit moves the pages on settings and customization to
a new directory configuration/.
It also splits the pages into meaningful pieces.

Change-Id: I65a3b0b688c8abc176662f86a0dbbfa3b64c4101