Updated openstack/openstack Project: openstack-infra/project-config  0518a27ee29bc63563c943780e7434b6adfa4c24
fix reference parsing in the unified publishing job for tagged releases

Fix an off-by-one error in the publishing job that made the job try to
set up doc/publish/tags/1.12.0 instead of doc/publish/1.12.0 (and fail
because the intermediate tags directory did not exist).

See
http://logs.openstack.org/7e/7e88422ffd84664bbdcf07c03428694d26ed3897/release/openstackdocstheme-docs-unified-ubuntu-xenial/26e912a/console.html#_2017-07-03_11_08_11_042419
for an example of a failure.

Change-Id: Iacbfdab106367e33f267665cd9463fc4789dcfba
Signed-off-by: Doug Hellmann <doug@doughellmann.com>