Updated openstack/openstack Project: openstack/openstack-manuals  14967e5ec3cc5881d925bf2bfb2071cce0dab804
update contributor guide based on changes after migration project

Update the instructions to indicate which guides are no longer
maintained by the docs team and how those docs are handled instead.

Explain the data input to the www-generator.py tool

Update the release instructions to explain how to update the templated
portion of the site.

Change-Id: I08bb228a6807bfcee0a6ab72ddaf229b455c9d1d
Signed-off-by: Doug Hellmann <doug@doughellmann.com>