Updated openstack/openstack Project: openstack-infra/project-config  57c5697dfc4d7ff20af18c012ebe5c3a973f2087
Remove old ironic doc jobs

Prior to the documentation migration, ironic had an
in-tree install-guide and in an in-tree api-ref.

With the documentation migration, ironic is stuck in
a catch-22 situation where we cannot move the content
in repo, without major re-organization of the content,
specifically with the installation guide and keep these
tests in place.

As such, I'm proposing this patch to remove the jobs
from ironic, as with top level TOC inclusion, they
should no longer be necessary.

Old install guide job will continue to run on pre-pike
releases to ensure they continue to be published.

Change-Id: I2709645b56035ac2f0971c52cc06c50af26b16f2