Updated openstack/openstack Project: openstack/openstackdocstheme  26f32fd0c1b72de83f4da6c7da2858d64dad4b06
Do not display title string of page local TOC

Some pages have no sub titles (the second level of titles or more).
In this case, the title string "Contents" of page local TOC is
completely meaningless and confusing to readers.

It is nice if this title string is shown only when needed.
This can be done by checking 'display_toc' variable in the template.

[1] https://docs.openstack.org/ocata/networking-guide/intro.html

Change-Id: I4803dd7712438eb2cd5fb6084e904bd8c0c98740