Updated openstack/openstack Project: openstack/heat  ad396377fd99b94e3d31fe140bd9ee01aba7da04
Resolve Macros when copying templates

During stack updates in the legacy path, we copy resource definitions
back and forth between templates. In the case where the definitions
contain macros (which in practice means the If macro for conditionals),
they may rely on external state (in practice, the conditional
definitions) that is not available in the template they're being copied
into (e.g. in the case of an If macro referencing a new condition).

This change means that when we copy a template, the macros get resolved
so that only the chosen path of the If macro is represented.

This resolves the issue where trying to signal a resource during an
update fails when one of the already-updated resources in the template
contains an If macro that refers to a condition definition that is
newly-added in the new template.

Change-Id: I6d08507f43b0fcc4c0b5e848e97fa26033d839b2
Closes-Bug: #1699463