Updated openstack/openstack Project: openstack-infra/project-config  51e0b583272719b2b377ba01f020363f7b1dd25d
Do not run install-guides for Pike anymore

With the doc-migration, we integrate all install-guides into the trees.
Thus stop building them for Pike. We need to run the job for Newton and
Ocata to support older builds.

Do this for all projects since all will do the migration, so there's no
need for multiple single changes.

Rearrange conditions so that logical units are together.

Change-Id: I6e22248dd7ed7736c8ab42848bc45d8e4238ccf1