Updated openstack/openstack Project: openstack/tempest  215e82b894563f30e9aaa6c665f9b5db95e5666c
Pause resource cleanup

When pause_teardown flag in tempest.conf is set to True a pdb breakpoint
is added to tearDown and tearDownClass methods in test.py.
This allows to pause cleaning resources process, so that, used resources
can be examined. Closer examination of used resources may lead to faster
debugging.

Change-Id: I09b7721e64cda161289f915d30888ec54bbed821