Updated openstack/openstack Project: openstack-infra/nodepool  c1db43b5154fae38a614f2cd069b63d5d1311dce
Remove duplicate python-jenkins code from nodepool

This is a redo of change I3852106ae6 which was reverted
in I53f1f86d3c6.  The reason why it didn't work previously
was because a bug to the create_node method was introduced
in python-jenkins while the nodepool change was sitting
around waiting to be approved.  That bug has now been fixed
in change I85fa43bc0b and a new release has been created with
the fix so this change should work now.

Change-Id: If54eaa9b4b9a0abd732f2cd22ac6b7ededc64855