Updated openstack/openstack Project: openstack/cinder  640b870f9c6a0fdc80c42f15af570f5d8df6779e
Revert "Don't use ignoreskipactivation for thin LVM"

Commit 8f5e10d7ed8eead221cccb6a0ccf1afea310b0a9 wrongly assumed that
the "ignoreskipactivation" flag is not needed when using thin LVM
provisioning. This is wrong.
Using thin LVM leads now to an error when the content will be copied:

ERROR cinder.volume.manager Command: sudo cinder-rootwrap /etc/cinder/
      rootwrap.conf dd if=/dev/mapper/cinder--volumes-_snapshot--UUID
      of=/dev/mapper/cinder--volumes-volume--UUID
      count=1073741824 bs=1M iflag=count_bytes conv=sparse

The reason is that the the snapshot is not activated and due to that,
no device node is available under /dev/mapper .

This reverts commit 8f5e10d7ed8eead221cccb6a0ccf1afea310b0a9 and makes
thin provisioning for LVM working again.

Change-Id: I788b297af92e11063b18c05a415b6f4d4d4bcd2c
Closes-Bug: #1702322
Related-Bug: #1701545