Updated openstack/openstack Project: openstack/cliff  0c378ba96fbbf7475a9e440417d018b42734c4d9
add --fit-width option to table formatter

Change the table formatter to not fit the terminal width by default. Add
a new --fit-width option to turn on that behavior. Width fitting is
implied if the --max-width argument is used. The old behavior can be
restored by setting CLIFF_FIT_WIDTH=1 in the shell environment.

Change-Id: I7b41b38ed25def416605d38c962aea55de842529
Signed-off-by: Doug Hellmann <doug@doughellmann.com>