Updated openstack/openstack Project: openstack-infra/storyboard-webclient  7b1f577d91cedcf7a23949b336ec6d9499d77a74
Move the parameter parsing into a service for reuse

This commit moves the code for parsing the parameters representing
search criteria in the URL into a separate service rather than the
StoryListController so that it can be reused elsewhere.

This commit also completes the implementation such that all useful
parameters can be parsed.

Change-Id: If2addfcab054a4e869d391dbda5f3845bbd701f6