Updated openstack/openstack Project: openstack/requirements  af21e8f67c4fdd65dd95303d097a06e7c9ca7b46
Remove astroid from global-requirements list

We dropped various linters in https://review.openstack.org/#/c/473094/.
This is a follow-up patch on this.

astroid is only used by pylint and both versions are tightly coupled.
If we force astroid===1.3.8, we cannot install newer pylint like 1.7.1.
It looks good to consider astroid as a linter.

Change-Id: Ib2e78f6a22b7d90a3304b13228c16d76982fbbef