Updated openstack/openstack Project: openstack/requirements  622aa9456c254a3181bf6bec3424a1f21c73993a
Update sphinxmark in u-c to resolve docs breakage

In 6dc6176 an update to Pillow to 4.2.0 was
included, which promptly broke sphinxmark 0.1.18.

Thankfully Doug Hellman already fixed sphinxmark
to work around the change in Pillow and there's a
new release.

We therefore update sphinxmark to the new release
so that we unbreak all the docs jobs which are
now failing.

Change-Id: If3268b1022e2350cbcee49235d0ca617eb7048c3