Updated openstack/openstack Project: openstack/oslo.config  ae80ab98842116364e746beb5684c92cc9f239cd
sphinxext: Add arguments for oslo.config:group directive

The 'oslo.config:group' directive was parsing content to extract the
name of the group. However, this had the side effect of picking up any
body text, resulting is super long titles. Correct this by adding a
single required argument to the directive.

This also has the side effect of not merging the content to a single
string, meaning formatting will be preserved as expected.

Change-Id: Ie2f011f4a2a0e4918b65bd2387ee2c41f4ae9ec2
Closes-Bug: #1702453