Updated openstack/openstack Project: openstack/horizon  b7a9754d3cdc4bc8a60e0b98dc003004a762bf72
Fix operation_log middleware tests

There seems to be a couple of errors on the operation_log
middleware tests that were being ignored on the tests
runnners.

On the first we were trying to access a local var from the
OperationLogMiddleware class to compare the default log
format. This is fixed by making it an attribute instead.

The second one was an icorrect use of the delattr function
as it was used as part of the request object, while its a
standalone function.

Change-Id: Ib05cfd1d6acb940a40c8bb1d1bdf3cdf837a0454