Updated openstack/openstack Project: openstack/openstack-manuals  3d2e93ec17396bd4be9083e121d1cb67bab11138
Redirect docs.o.o/repo always to /latest

We need to redirect for all links, not only for services with service
types.

Currently https://docs.openstack.org/horizon/ and
https://docs.openstack.org/openstack-ansible/ and
https://docs.openstack.org/openstackdocstheme/ do not redirect.

This change is not complete, it only covers horizon. The other two types
of projects (not listed in latest.yaml) are not covered.

Change-Id: I8df6f734b97bbc5d3adf0c040a59031e44f46542